format version data type in db (int to bigint)

+++ src/main/resources/db/migration/postgresql/V10__init_activity_participant.sql
@@ -1,9 +1,9 @@
 CREATE TABLE activity_participant
 (
-    id               UUID         NOT NULL DEFAULT uuid_generate_v4(),
-    user_id          UUID         NOT NULL,
-    post_id          UUID         NOT NULL,
-    version          INTEGER      NOT NULL,
+    id               UUID                        NOT NULL DEFAULT uuid_generate_v4(),
+    user_id          UUID                        NOT NULL,
+    post_id          UUID                        NOT NULL,
+    version          BIGINT                      NOT NULL,
     created_at       TIMESTAMP WITHOUT TIME ZONE NOT NULL,
     last_modified_at TIMESTAMP WITHOUT TIME ZONE NOT NULL,
     PRIMARY KEY (id),
