+++ src/main/resources/db/migration/postgresql/V2__init_university.sql
@@ -5,7 +5,7 @@ CREATE TABLE university
     description      VARCHAR(255),
     profile_img_id   UUID,
     admin_id         UUID                        NOT NULL,
-    version          INTEGER                     NOT NULL,
+    version          BIGINT                      NOT NULL,
     created_at       TIMESTAMP WITHOUT TIME ZONE NOT NULL,
     last_modified_at TIMESTAMP WITHOUT TIME ZONE NOT NULL,
     PRIMARY KEY (id),
